MicrojetDetection (Python)/scripts/requirements.txt
PyQt5_sip == 12.12.1
numpy == 1.23.5
xlwt == 1.3.0
matplotlib == 3.7.2
opencv_python == 4.6.0.66
